finalizing turn system and dividing gamepage into components
